requirements.txt
mysql-connector-python==8.0.16
mysql-connector==2.2.9
